Add files via upload
This Power BI dashboard helps users understand the profitability of a fictional retail company by visualizing its sales and profit data from the Superstore dataset. The visualizations include bar charts, line charts, and tables that can be filtered by category, sub-category, region, and year. The dashboard also includes calculated measures such as gross profit margin, net profit, and profit percentage. Users can interact with the visualizations to explore the data in different ways and identify areas where the company could improve its profitability. This dashboard can be used as a reference for learning how to build effective visualizations in Power BI or as a template for creating similar dashboards with your own data.


Tools Used :- 

Power BI Desktop
Microsoft Excel

Visualizations:-

In this project, we have created various visualizations to gain insights into the sales data of the superstore, including:

 Profit by Category
 Profit by Segments
 Profit by Region
 Profit by Year and Quarter 
 Profit by Sub Category 
Total Profit, Sales and Quantity 

Conclusion:-

In conclusion, the Superstore Profit Visualization on Power BI is a comprehensive dashboard that presents an in-depth analysis of sales and profit data from the Superstore dataset. It provides users with valuable insights into the company's profitability by category, region, and year, and can be used as a reference for learning how to build effective visualizations in Power BI. With its interactive and customizable features, this dashboard serves as a useful tool for exploring and analyzing data and can be adapted to fit the specific needs of different businesses and industries.